+++ tests/Amadeus/Client/Struct/Fare/MasterPricerTravelBoardSearchTest.php
@@ -1141,4 +1141,74 @@ class MasterPricerTravelBoardSearchTest extends BaseTestCase
         $this->assertEmpty($message->fareFamilies[1]->otherPossibleCriteria);
         $this->assertEmpty($message->fareFamilies[1]->fareFamilySegment);
     }
+
+    /**
+     * 5.32 Operation: 02.32 Flight Option - Progressive legs
+     */
+    public function testCanMakeMessageWithProgressiveLegs()
+    {
+        $opt = new FareMasterPricerTbSearch([
+            'nrOfRequestedPassengers' => 1,
+            'passengers' => [
+                new MPPassenger([
+                    'type' => MPPassenger::TYPE_ADULT,
+                    'count' => 1
+                ])
+            ],
+            'flightOptions' => [
+                FareMasterPricerTbSearch::FLIGHTOPT_PUBLISHED
+            ],
+            'itinerary' => [
+                new MPItinerary([
+                    'departureLocation' => new MPLocation(['city' => 'DEN']),
+                    'arrivalLocation' => new MPLocation(['city' => 'LAX']),
+                    'date' => new MPDate([
+                        'dateTime' => new \DateTime('2015-12-11T00:00:00+0000', new \DateTimeZone('UTC'))
+                    ])
+                ]),
+                new MPItinerary([
+                    'departureLocation' => new MPLocation(['city' => 'LAX']),
+                    'arrivalLocation' => new MPLocation(['city' => 'BOS']),
+                    'date' => new MPDate([
+                        'dateTime' => new \DateTime('2015-12-18T00:00:00+0000', new \DateTimeZone('UTC'))
+                    ])
+                ])
+            ],
+            'progressiveLegsMin' => 0,
+            'progressiveLegsMax' => 1
+        ]);
+
+        $message = new MasterPricerTravelBoardSearch($opt);
+
+        $this->assertCount(1, $message->numberOfUnit->unitNumberDetail);
+        $this->assertEquals(1, $message->numberOfUnit->unitNumberDetail[0]->numberOfUnits);
+        $this->assertEquals(UnitNumberDetail::TYPE_PASS, $message->numberOfUnit->unitNumberDetail[0]->typeOfUnit);
+
+        $this->assertCount(1, $message->fareOptions->pricingTickInfo->pricingTicketing->priceType);
+        $this->assertEquals(
+            [
+                PricingTicketing::PRICETYPE_PUBLISHEDFARES
+            ],
+            $message->fareOptions->pricingTickInfo->pricingTicketing->priceType
+        );
+
+        $this->assertCount(2, $message->itinerary);
+        $this->assertEquals('DEN', $message->itinerary[0]->departureLocalization->departurePoint->locationId);
+        $this->assertEquals('LAX', $message->itinerary[0]->arrivalLocalization ->arrivalPointDetails->locationId);
+        $this->assertEquals('111215', $message->itinerary[0]->timeDetails->firstDateTimeDetail->date);
+        $this->assertNull($message->itinerary[0]->timeDetails->firstDateTimeDetail->time);
+        $this->assertNull($message->itinerary[0]->timeDetails->firstDateTimeDetail->timeWindow);
+        $this->assertNull($message->itinerary[0]->timeDetails->firstDateTimeDetail->timeQualifier);
+        $this->assertNull($message->itinerary[0]->timeDetails->rangeOfDate);
+
+        $this->assertEquals('LAX', $message->itinerary[1]->departureLocalization->departurePoint->locationId);
+        $this->assertEquals('BOS', $message->itinerary[1]->arrivalLocalization ->arrivalPointDetails->locationId);
+        $this->assertEquals('181215', $message->itinerary[1]->timeDetails->firstDateTimeDetail->date);
+        $this->assertNull($message->itinerary[1]->timeDetails->firstDateTimeDetail->time);
+        $this->assertNull($message->itinerary[1]->timeDetails->firstDateTimeDetail->timeWindow);
+        $this->assertNull($message->itinerary[1]->timeDetails->firstDateTimeDetail->timeQualifier);
+        $this->assertNull($message->itinerary[1]->timeDetails->rangeOfDate);
+
+        $this->assertCount(2, $message->travelFlightInfo->unitNumberDetail);
+    }
 }